backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-dotenv
django-cors-headers
Pillow
google-auth-oauthlib
google-api-python-client
requests
google-generativeai
pytesseract
PyPDF2
python-docx
pandas
openpyxl


pytesseract


pillow


pdfplumber


python-docx


openpyxl
